+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to the
 ### Added
 - More unit tests
 - `verifyJournal` and `saveJournal` functions
+- Functions to compute the account flow and balance
+- Function `failedAssertion` to find the failed balance assertions
 
 ## 0.1.0 - 2024-12-29
 ### Added